Port model_base to TF2
This patch ports most of the functionality of model_base to TF2 along
with all the tests. Notably left out is run_continuous_evaluation as it
is tested separately. This will be fixed in a future patch where the
tests for that are updated as well.

Pull Request: https://github.com/google/gematria/pull/280