Make GraphNetworkLayer subclass from tf.Module
We cannot make the dataclass frozen when doing this, but this enables TF
to automatically find trainable variables within GraphNetworkLayer
objects which means we can get rid of the somewhat hacky
_get_trainable_variables function that subclasses were supposed to
override. This successfully trains models that would otherwise fail to
converge if none of the graph layers were trainable.

This closes #323.

Reviewers: ondrasej

Pull Request: https://github.com/google/gematria/pull/337