Annotate training step with tf.function
This patch annotates the entire training step with tf.function. This
provides little performance improvement on the CPU side, but should
provide a large benefit on the GPU side, where we can also take
advantage of XLA compilation by setting jit_compile=True. This is left
for a future patch as we need to figure out a good way to turn it on and
off.

Pull Request: https://github.com/google/gematria/pull/345